+++ tests/dir-statusline/mock-diagnostics.lua
@@ -1,9 +1,3 @@
-vim.lsp.buf_get_clients = function() return { 'mock client' } end
-if vim.fn.has('nvim-0.8') == 1 then
-  vim.api.nvim_exec_autocmds('LspAttach', { data = { client_id = 1 } })
-  _G.detach_lsp = function() vim.api.nvim_exec_autocmds('LspDetach', { data = { client_id = 1 } }) end
-end
-
 vim.diagnostic.get = function(_, _)
   local s = vim.diagnostic.severity
   return {